Limit tx history
