+++ packages/extension/src/background/accounDeployAction.ts
@@ -1,4 +1,4 @@
 import { ExtQueueItem } from "../shared/actionQueue/types"
 import { BaseWalletAccount } from "../shared/wallet.model"
 import { BackgroundService } from "./background"
-import { addTransaction } from "./transactions/store"
+import { addTransaction } from "../shared/transactions/store"
