+++ packages/extension/src/background/transactions/badgeText.ts
@@ -8,7 +8,7 @@ import { Transaction } from "../../shared/transactions"
 import { BaseWalletAccount } from "../../shared/wallet.model"
 import { accountsEqual } from "../../shared/wallet.service"
 import { walletStore } from "../../shared/wallet/walletStore"
-import { transactionsStore } from "./store"
+import { transactionsStore } from "../../shared/transactions/store"
 
 // selects transactions that are pending and match the provided account
